tools: widen the codegen gate to the property, not gcc's loop shape
Built with clang 19.1, make test stopped before running anything:
fzn_ct_memeq compiles to two conditional branches there and the gate
demanded exactly one.

The gate's own message says to read the disassembly and decide rather than
relax it, so that is what happened. clang's two are a zero-length guard,
test %rdx,%rdx followed by je, and the back edge, cmp %rax,%rdx followed
by jne. Both are on the length and neither touches the data. gcc emits a
top-tested loop, one conditional plus an unconditional back edge; clang
emits a rotated one. Same control flow, different shape, and the gate was
pinned to one compiler's.

The bound is one or two now, and it is measured rather than generous.
Across gcc 14.2 and clang 19.1, real is 1/1/1 and 2/1/1 in branches, sets
and returns; the early exit is 2/0/2 and 4/1/2; the no-volatile variant is
1/1/1 and 2/1/2. gcc's early exit is the only sabotage that fits inside
the relaxed branch bound, and it is still refused by the missing
conditional set and the second return. All six sabotages are refused at
both compilers, and none survives on the branch count alone, which is what
makes widening it cost nothing.

The suite then passes under clang, 23 binaries and zero failures, with one
diagnostic gcc does not give: situ_fzn_kind_t narrowing to uint8_t in
wire/seal.c. The values are 0x00 to 0x03 so nothing is lost, and the cast
is explicit now -- two compilers disagreeing about whether a conversion is
worth mentioning is worth one line in the source.

+++ wire/seal.c
@@ -95,7 +95,11 @@ fzn_seal_err_t fzn_seal_open(uint8_t *frame, size_t frame_len,
 	out->msg = situ_fzn_head_msg_get(hv);
 	out->index = situ_fzn_head_index_get(hv);
 	out->chunks = situ_fzn_head_chunks_get(hv);
-	out->kind = situ_fzn_head_kind_get(hv);
+	/* Cast explicit: the accessor returns the schema's enum and `kind` is a
+	 * byte, which clang reports as a narrowing and gcc does not. The values
+	 * are 0x00 to 0x03 so nothing is lost, and saying so in the source beats
+	 * two compilers disagreeing about whether it is worth mentioning. */
+	out->kind = (uint8_t)situ_fzn_head_kind_get(hv);
 
 	return FZN_SEAL_OK;
 }
